Added PG articles to structure
Added the Publishing Group section to the Courses and Activities landing page. Added the articles to the section, to the Courses and Activities A-Z section, and to the AZ landing page. Replaced the ...add SCORM modules article with the What is a PG article on the Home page in the TMS Administrators / Courses and Activities area.